модуль print_mixin
