+++ src/main.py
@@ -2,18 +2,7 @@ from abc import ABC, abstractmethod
 from typing import Any
 from src.base_product import BaseProduct
 from src.base_category_order import BaseCategoryOrder
-
-
-class PrintMixin:
-    """Класс-миксин для печати в консоль информации"""
-
-    def __init__(self):
-        """инициализация миксина"""
-        print(repr(self))
-
-    def __repr__(self) -> str:
-        """вывод сообщения об объекте"""
-        return f"{self.__class__.__name__}({self.name},{self.description},{self.price},{self.quantity})"
+from src.print_mixin import PrintMixin
 
 
 class Product(BaseProduct, PrintMixin):
